Manpage + CHANGELOGs
Also, since there is some change of someone miskeying --virt=yed or something, and getting a False
value (which means reprovision the main OS), the default for --virt parameters is getting changed to assume --virt implies True when --virt is used.  We really don't want a typo erasing an important server.