Lots of work towards profile inheritance.  This works in the UI now, with some rough edges (like listing the tree).
cobbler profile add --name=profile2 --inherit=profile1 --otherparameters=...
cobbler profile edit --name=profile2 --stillmoreparamters=...

Data is interleaved for hashes, combined for arrays, and overriden for scalar
values.  This was heavily inspired by Will-It-Blend, and in this implementation
it all blends.

Implementation notes -- Updating a parent profile doesn't apply changes
to the child objects until a sync, so this seems like a good upgrade
for a future commit.  Also, the children mapping that makes this
possible needs some tweaks because they may load out of order,
in which case "cobbler list" can't render a full
tree.  There are various approaches to deal with this and it should be
a (relatively) easy change.