Keep track of depth of cobbler objects such that a pseudo-topological sort can be done at deserialization time.  The result of this is that full graph listing information can be reconstructed at time of config loading (up and down links), while only having to store the up links.
This preserves the existing config file format while allowing for arbitrary inheritance and
a reasonable measure of hand-editability.

If changing profile relationships by hand, the cached depth info may be wrong,
so some way to automatically resolve this could potentially be doable,
but it's such a distinct corner case that I don't deem it neccessary at this point.