Ignore non-present network interfaces in post_install_network_config
If during execution of this snippit $IFNAME is empty, the the script
will end up executing:

grep $IFNAME /etc/modprobe.conf
which will be:
grep /etc/modprobe.conf

this never will return and results in a hung kickstart.  Now the only
reason IFNAME would be null is if at provisioning time, an interface
specified in the system record wasn't present.  This might be an 'error
case', but it would be more ideal to simply ignore the interface than to
hang.

Here's the patch:

index e6839f7..7191087 100644