Bugfix: run createrepo on partial yum mirrors
Though the comment removed here suggests that the original intent was to skip
running createrepo(8) on full mirrors created with reposync(1), the check was
reversed, and createrepo was in fact skipped for partial mirrors created with
yumdownloader(1). As it happens the comment is wrong, and we actually have to
run createrepo for *all* local yum mirrors.