A smart tftp server, and a module to manage it
* Can be run standalone or from xinetd
* handles standard extensions (large blocksize, tsize)
* Looks up hosts via IP address via cobbler (XMLRPC)
* Uses the host's template_files materialized property, if applicable.

For example, if the host's template_files property included
/var/www/cobbler/ks_mirror/foo-bar/template/blah=/boot/blah

and the client request the file "/boot/blah", then it will get the file from
template/blah, and that file will be materialized as a template (with the
hosts's flattened data structure as the data source) for the requesting
host.

A new management module: tftpd_py.
  It replaces in_tftpd, and basically does a lot less: because the
  tftpd server now does more.