Support installing FreeBSD without an IP address set in the host record.
FreeBSD's pxeboot does not do the sort of filename templating that pxelinux.0
does, so you can not easily look up the host by mac_address the way that you
can from pxelinux.0 requests.  However, since the freebsd install process does
go through pxelinux.0 (to chain bootloaders), I can catch the pxelinux.0
request for pxelinux.cfg/<mac_address>, and use that MAC to look up the host,
then cache the ip->hostname mapping for future reference.

This commit adds that map.  The tftpd server does not use it by default,
you need the "-c" option.  Data is only cached for 300 seconds.  Because
the cache relies on a sequence of events all being handled by the same
process, it's more fragile than I'd like: if the tftpd server dies
in the middle (slow server, timeouts), if the client is extremely slow
and outlives the cache ...