Wil Cooley <wcooley@nakedape.cc>
This patch corrects a 'No such file or directory' error from wget in the
case where "mirror locally" is false but "keep updated" is true and the
repos were created with an older version of Cobbler.

I don't know if this combination of options actually makes sense but
they're accepted as valid and this change of behavior causes errors
(bogus or not) in the reposync output. It is also logical at least
within the context of the existing code; the wget of repomd.xml that
starts around line 391 does not require 'repo.mirror_locally', so the
os.makedirs shouldn't either.

Wil